--- app/src/main/java/com/example/munchkin/Karte/RustungCards.java
@@ -1,5 +1,0 @@
-package com.example.munchkin.Karte;
-
-public interface RustungCards extends Karte {
-    void addBonus(int currentLvl);
-}
